+++ android/src/main/kotlin/com/squareup/square_mobile_payments_sdk/SquareMobilePaymentsSdkPlugin.kt
@@ -92,6 +92,9 @@ class SquareMobilePaymentsSdkPlugin : FlutterPlugin, MethodCallHandler, StreamHa
       "removeReaderChangedCallback" -> ReaderModule.removeReaderChangedCallback(result)
       "pairReader" -> ReaderModule.pairReader(result)
       "stopPairing" -> ReaderModule.stopPairing(result)
+      "readerSettings" -> ReaderModule.readerSettings(result)
+      "closeSettings" -> SettingsModule.closeSettings(result)
+      "isShowingSettings" -> SettingsModule.isShowingSettings(result)
       else -> result.notImplemented()
     }
   }